경로를 따라 움직이는 비히클 구현

+++ Code/PrevTestLab_Win32/PrevTestLab_Win32/CObjectFuncTemplates.h
@@ -108,8 +108,8 @@ inline bool LineIntersection2D(CVector2D   A,
 		return false;
 	}
 
-	double r = rTop / rBot;
-	double s = sTop / sBot;
+	float r = rTop / rBot;
+	float s = sTop / sBot;
 
 	if ((r > 0) && (r < 1) && (s > 0) && (s < 1))
 	{
